Updated README.md and Organize the folder
Update the README.md file and add How to contribute to this repository and organize the folder
