+++ Question 1/ArrayDuplicate.cpp
