+++ Question 1/C++Solution.cpp
