+++ Question 1/Duplicate in an array.js
