+++ Question 1/SolutionQ1.java
