+++ Question 1/ans.cpp
