+++ Question 1/answer.java
