+++ Question 1/cpp_ans.cpp
