+++ Question 2/C++Solution.cpp
