+++ Question 2/answer.java
