+++ Question 2/solution.cpp
