+++ Question 3/C++Solution.cpp
