+++ Question 3/boundedMatrixTraversal.py
