+++ Question 3/bounded_traversal_java.java
